Migrate the big functions to their separate files
