+++ indexer/main/src/main.rs
@@ -1,6 +1,6 @@
 // Massbit dependencies
-use index_manager::types::IndexManager;
 use logger::core::init_logger;
+use index_manager::core::IndexManager;
 
 #[tokio::main]
 async fn main() {
